DLP: Add feature to disable DLP policy parsing and applying.
Adding a feature flag that is disabled by default as the behavior even
though it is anyway gated on the presence of the policy.
For now we'll enable it via Finch flag configuration for all users.
However, we need to have ability to turn the feature off on older builds
later when the policy will be released to public - otherwise these
devices with older build will have unwanted experience.

Bug: 1131067
Test: New unit test added.
Change-Id: Ia29f053f4128d67a2976e78a573f1ccda998f7c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2428898
Reviewed-by: Nikita Podguzov <nikitapodguzov@chromium.org>
Reviewed-by: Aya Elsayed <ayaelattar@chromium.org>
Commit-Queue: Sergey Poromov <poromov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#811977}