[MBI]: Migrate RenderThread::GetRemoteRouteProvider() calls through ASG
This CL migrates all calls to RT::GetRemoteRouteProvider() through ASG,
so that it is the sole mojom::RouteProvider* in the renderer. This CL
further restricts the usage of RT::GetRemoteRouteProvider to ASG via
util::PassKey.

For now, we still delegate to the RenderThread implementation to provide
AssociatedInterfaceProvider routes. A subsequent CL will make
AgentSchedulingGroup a fully-fledged RouteProvider, and we'll remove the
RenderThread::GetRemoteRouteProvider public API then.

Bug: 1132901
Change-Id: Idbd9311c777736d0a478da7a66718baa4e723275
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2433590
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Tal Pressman <talp@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Commit-Queue: Dominic Farolino <dom@chromium.org>
Cr-Commit-Position: refs/heads/master@{#812026}