[OsSettingsLanguages] Update Keyboard section link to input
View when flag is on: http://screen/4hzTLaxE66FMnce
Normally, this would link to the full Languages and Inputs page.
But with Languages and Inputs being in separate page, UX thinks
linking to Inputs directly is more suitable as users are in Keyboard
section.

Bug: 1113439
Test: Manual test
Change-Id: I6087e7ff4d974daf6648b2bc45638e5f27805815
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2437717
Reviewed-by: Regan Hsu <hsuregan@chromium.org>
Commit-Queue: My Nguyen <myy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#812061}