[iOS] Remove CHROME_CONNECTED cookie after displaying sign-in promo.
Currently Chrome stores whether it should show a sign-in promo on
accounts.google.com in the CHROME_CONNECTED cookie. Once the promo is
displayed it no longer requires this information and we should remove
the cookie.

This will not affect existing functionality since the CHROME_CONNECTED
cookie is used for signed-in users and the promo will only be displayed
to signed out users.

Bug: 1125631
Change-Id: I504974aaebf8e7e71a596342726c5f5df0f189c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2440596
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Nohemi Fernandez <fernandex@chromium.org>
Cr-Commit-Position: refs/heads/master@{#812078}