Reland "Revert "[ios] Add EG test to chrome://management page""
Reland https://chromium-review.googlesource.com/c/chromium/src/+/2432316 with fix.

The third patchset is the reverted cl.

The issue is caused because the link navigates to an external page,
but the test can't always load a external page, so the trybot fails
randomly. Fixed by only checking if the link can be tapped.

Change-Id: I932e8976b6b99236003420cd87e21f6c339f2376
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2436774
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Tina Wang <tinazwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#812095}