Support Bloom in |AssistantInteractionController|
This means the prototype now works, and the bloom response can be seen
in the Assistant UI (as a very unreadable JSON string).

I changed the |BloomInteractionObserver| into a |BloomUiDelegate|,
as that seemed a better approach to keep the Bloom code independent of
ash.

Bug: b/165356952
Test: Manually deployed
Change-Id: I4e1fd1acd0644c97b3e5e91131d4e593353b1e1c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2390944
Auto-Submit: Jeroen Dhollander <jeroendh@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Commit-Queue: Jeroen Dhollander <jeroendh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#812103}