Add the unlisted feature tag for policy templates
The unlisted tag is designed for policies that are not controlled by
admin directly but set by the device management server directly. Because
of that, the tag can only be used by cloud_only policies.

The unlisted policies will be hidden from all documentation and template
generators. On the  Chromium side, this is achieved by the cloud_only
tag.

Bug: 1117637
Change-Id: I8e6e2bdb681a3e68897ec9e41f285805dc2e37d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2436697
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Commit-Queue: Owen Min <zmin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#812112}