Record if the navigation's initiator is available.
In NavigationRequest::BeginNavigation(), records whether the
navigation's initiator can be found using its ID or not.

We suspect not finding it is a very rare event. This histogram will help
us figure out what kind of solution should be used to support
PolicyContainer for this edge case.

Bug:1129416

Change-Id: I501bf0150935679667d5bd2a4cc351a70b6653ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2416450
Reviewed-by: Camille Lamy <clamy@chromium.org>
Reviewed-by: Pâris Meuleman <pmeuleman@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Antonio Sartori <antoniosartori@chromium.org>
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#812124}