capture_mode: Styling and location of dimensions label.
The dimensions label for capture mode shows user the dimensions of the
current capture region. Currently it is not styled, nor does it position
itself reactively.

This CL styles the label and also makes it reposition itself depending
on the position of the capture region.

Tests: manual + added tests.
Bug: 1131118
Change-Id: I5bd0283a2d2057ba76ba75f3d76c2856f261d89b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2429588
Commit-Queue: Jeremy Chinsen <chinsenj@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Sammie Quon <sammiequon@chromium.org>
Cr-Commit-Position: refs/heads/master@{#812207}