Add font matching metrics recording to OffscreenFontSelector
The FontMatchingMetrics class records UKMs about the font matching on
the web. This will be used for the identifiability study to understand
the privacy leak that font matching can cause. Currently, these metrics
are only recorded for matching in the CssFontSelector (i.e. occurring on
documents).

This cl allows construction of a FontMatchingMetrics object on
WorkerGlobalScopes and adds reporting from the OffscreenFontSelector.
This ensures that the font matching occurring on workers is recorded, so
that any resulting privacy leak can be studied.

Bug: 1131928, 1025945
Change-Id: Ic8794813453d8a715071cfa71f1c4fed7444f58e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2197604
Reviewed-by: Asanka Herath <asanka@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Commit-Queue: Alex Turner <alexmt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#812223}