[Sampling profiler] Factor out platform config 4/4
Fourth in a series factoring the ThreadProfiler platform specific
configuration state from the code that takes action on the state.
Defines a GetChildProcessEnableFraction function that returns how
frequently the profiler should be enabled within a given process.

The end goal is to reduce the configuration complexity, to support
enabling per-thread on Android.

Bug: 1129939
Change-Id: I936c671cd30f1750cf87d6ec8ce9aee45c051be4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2424672
Commit-Queue: Mike Wittman <wittman@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Etienne Pierre-Doray <etiennep@chromium.org>
Cr-Commit-Position: refs/heads/master@{#812248}