MediaRouter: move incognito shutdown code to //chrome
This is necessary for Chrome because a normal profile and an incognito
profile use the same MediaRouter. That object should be notified when
the incognito profile shuts down.

This is not necessary for WebLayer because OTR contexts have their
own MediaRouter (the OTR context is not linked to a normal/base
context).

Bug: 1057100
Change-Id: I9c4426a4219bd971c6690b4d2712e5d39ec3f91e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2438946
Reviewed-by: Takumi Fujimoto <takumif@chromium.org>
Commit-Queue: Evan Stade <estade@chromium.org>
Cr-Commit-Position: refs/heads/master@{#812250}