ozone: evdev: classify pointing sticks separately
In order to add a separate settings section for pointing sticks (a.k.a.
TrackPoints), we need to be able to distinguish them from mice. This CL
uses the `INPUT_PROP_POINTING_STICK` evdev property to identify them,
then uses booleans to report whether mice and pointing sticks are
present. The `InputController` interface still counts them as mice for
now (to retain the existing behaviour until the next CL is submitted).

We take this approach (as opposed to creating separate lists, like we do
for touchpads) because `DeviceHotplugEventObserver` only has the one
category, and we can't add another one there as it's also used in the
X11 stack, which doesn't distinguish pointing sticks.

Bug: 1114828
Test: check mouse settings still appear when a mouse is connected

Change-Id: I7018f78f2af26e61540438385a3c6c9ced15265b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2423405
Reviewed-by: Michael Spang <spang@chromium.org>
Commit-Queue: Harry Cutts <hcutts@chromium.org>
Cr-Commit-Position: refs/heads/master@{#812256}