[CrOS PhoneHub] Get Browser tab metadata including favicons.
Creates a helper class that gets the metadata needed to return
a vector of the latest visited BrowserTabMetadata to the caller.

Creates a fake version that can be used for unit tests of
callers to this class.

Note the class requires a FaviconService to be passed in the ctor.

Bug: 1106937
Change-Id: Ibe669cc337190025c59d536f36450c054731e14f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2424513
Commit-Queue: Regan Hsu <hsuregan@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#812274}