Disable CSD pings for interstitial click throughs
This CL removes the functionality for sending a CSD ping when an
interstitial was previously shown on the page. This is the minimal CL
to do so, and is intended to be merged.

Future CLs will clean up useless code introduced by this CL.

Bug: 1132060
Change-Id: I9f3e163268254cf79eee60b30b2d3da025293abc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2429809
Reviewed-by: Bettina Dea <bdea@chromium.org>
Reviewed-by: Varun Khaneja <vakh@chromium.org>
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Cr-Commit-Position: refs/heads/master@{#812289}