Assume float and half float formats are renderable on Android
except for RGB32F.

This is true since ES3 and most of Android devcies are on ES3+,
and histogram data confirmed this: 100% renderable for all formats,
except RGB32F, which is 99.99%+ unrenderable.

Avoid runtime check - this is a minor optimization in validating
command buffer on Android.

BUG=1127837
TEST=bots
R=kbr@chromium.org

Change-Id: I9ac1c0c87ae197d6883e9167068bd81fc4e0d844
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2438313
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Commit-Queue: Zhenyao Mo <zmo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#812305}