[devtools] Reduce performance overhead of message dispatch in the front-end.
rebase CL 2044170 part 1:
Adds a new `FrameNavigationControl::JavaScriptMethodExecuteRequest`
method. Previously every message was turned into an object literal in a new
script and sent to the renderer process executing the
front-end, which was very costly and leads to a lot of unnecessary
script executing in the V8 engine.

Bug: chromium:1029427
Change-Id: I9e803ea9d32ce333c4eea1d2cb06f1b2eaffaf02
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2431864
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Maksim Sadym <sadym@chromium.org>
Cr-Commit-Position: refs/heads/master@{#812308}