[WebOTP] Record pending origin count in SmsFetcher
We currently use a map to store "origin, subscribers" pairs. Multiple
pending origins may be rare in practice. e.g. a user visits origin A,
while waiting for an SMS by WebOTP, they visit origin B and wait for
another SMS.

We want to measure the frequency of the multiple pending origins
occurrenceto to evaluate other metrics. e.g. when an sms from one origin
cannot be parsed, or the prompt is canceled/timeout, The simplest most
practical approach is to report it on all existing pending origins which
could lead to over-counting if multiple pending origins are a common
occurrence.

Bug: 1127397
Change-Id: Ife092d8e038b96977b34740854e5ed0a690083af
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2426974
Commit-Queue: Yi Gu <yigu@chromium.org>
Reviewed-by: Sam Goto <goto@chromium.org>
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Reviewed-by: Majid Valipour <majidvp@chromium.org>
Cr-Commit-Position: refs/heads/master@{#812311}