Fix blob url loading in <webview> in guest mode.
As it turns out Chrome OS guest mode is violating the assumption
<webview> made that a chrome app's storage partition is always on disk.
Because of this, before the change that added the blob url fallback
mechanism a <webview> with no explicit partition name set would
(incorrectly) share the storage partition with the app that embeds it.
Adding the blob url fallback flag fixed that (by making them separate
configs only differing in the fallback flag), but the fallback mechanism
also assumed that a chrome app's storage partition always was on disk.

This CL fixes loading blob URLs in a <webview> in guest mode by changing
StoragePartitionConfig to store a tri-state enum rather than a simple
bool for its blob url fallback mode. This way we can support fallback
regardless of if the chrome app's storage is in memory or on disk.

Bug: 1128330
Change-Id: I518cd5d1e89491eaf1cc6e2ed9ddb5cd3f0b9e87
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2439110
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Lucas Gadani <lfg@chromium.org>
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Reviewed-by: Aaron Colwell <acolwell@chromium.org>
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#812313}